add new FUnctions

+++ src/components/alerts/AlertAuth.tsx
@@ -3,17 +3,19 @@ import logo from '../../assets/logo.svg';
 import { BsFillCheckSquareFill } from 'react-icons/bs';
 import { MdError } from 'react-icons/md';
 import { useSelector, useDispatch } from 'react-redux';
-import { AppStore } from '../../redux/store'
-import { setearAlertRemover } from '../../redux/states/alertAuth'
+import { AppStore } from '../../redux/store';
+import { setearAlertRemover } from '../../redux/states/alertAuth';
 export default function AlertAuth() {
     const alertAuth = useSelector((state:AppStore)=>state.alertAuth);
-    const dispatch = useDispatch()
+    const dispatch = useDispatch();
     useEffect(()=>{
-        setTimeout(()=>{
-            dispatch(setearAlertRemover());
-        },4000)    
+        if(alertAuth.render){
+            setTimeout(()=>{
+                dispatch(setearAlertRemover());
+            },4000)  
+        }
     },[alertAuth.render])
-  return  alertAuth.render && (
+  return  alertAuth.render ? (
 
     <div className='containerFather'>
         <div className="containerCard">
@@ -24,7 +26,7 @@ export default function AlertAuth() {
                 {alertAuth.alert[0] !== "error" ? 
                 <BsFillCheckSquareFill color='green' size="3.5rem"/> :
                 <MdError color='red' size="3.5rem"/> 
-                }
+            }
             </div>
             <div className='containerDescription'>
                 {
@@ -33,6 +35,7 @@ export default function AlertAuth() {
             </div>
         </div>
     </div>
-    )  
-  
+    )
+    :
+    null
 }
